+++ NEWS.md
@@ -3,6 +3,7 @@
 * Changed default value of forbidden codons to ""
 * Make result plots more flexible
 * Propagate information about the number of mutated bases/codons in `summarizeExperiment()`
+* Added function to calculate nearest string distances
 
 # mutscan 0.2.22